requirements.txt
# MetaGaze - Gaze Estimation System
# Python Dependencies
# ================================
# Install: pip install -r requirements.txt

# --- Web Framework ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=11.0

# --- Deep Learning ---
torch>=2.0.0
torchvision>=0.15.0

# --- Computer Vision ---
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0

# --- Scientific Computing ---
numpy>=1.24.0

# --- Data Handling ---
h5py>=3.9.0

# Note: Install frontend dependencies separately with:
# cd frontend && npm install
